Operational Search Button added to the action bar.
An indeterminate ProgressBar added while fetching the books from the OpenLibrary API.

A Detail View is added to display more details of a book when selected from the Books List.

A share icon is added as well on the detail view screen to allow users to share a book.

ic_launcher change.

Signed-off-by: Plaisir93 <plaisirpierrecharles@gmail.com>